Enable -Wall -Werror in both build systems
Promote all warnings to errors. Fix all remaining issues:

- Fix -Wliteral-suffix: add spaces around AUTOCOLUMN_SEP and
  SQLITE_TEMP_FILE_PREFIX macro concatenation
- Fix -Wvolatile: move volatile from parameter to local variable
  in kahanBabuskaNeumaierStep (C++20 deprecates volatile params)
- Fix -Wparentheses: remove unnecessary parens in azHelp declaration
- Fix -Wmaybe-uninitialized: initialize ePtrmapType and iPtrmapParent
- Fix -Wsign-compare: cast sizeof() to int in loop comparisons
- Fix -Wunused-variable/-Wunused-but-set-variable: suppress with (void)
- Fix -Wregister: remove register keyword in test_md5.cpp
- Fix volatile increment: seenInterrupt = seenInterrupt + 1
- Fix global.cpp: remove unnecessary extern from non-const pointer
  definitions; preserve extern on const array declarations in
  amalgamation generator
- Fix sqlar.cpp sign-compare with explicit cast

Build tools (jimsh, lemon, mkkeywordhash, mksourceid) are compiled
as C and excluded from -Werror.

Test results: 332,259/332,259 pass (0 errors)

+++ ext/expert/test_expert.cpp
@@ -202,7 +202,7 @@ int TestExpert_Init(Tcl_Interp *interp){
   };
   int i;
 
-  for(i=0; i<sizeof(aCmd)/sizeof(struct Cmd); i++){
+  for(i=0; i<(int)(sizeof(aCmd)/sizeof(struct Cmd)); i++){
     struct Cmd *p = &aCmd[i];
     Tcl_CreateObjCommand(interp, p->zCmd, p->xProc, 0, 0);
   }
