+++ ext/rbu/test_rbu.cpp
@@ -418,7 +418,7 @@ int SqliteRbu_Init(Tcl_Interp *interp){
     { "sqlite3rbu_internal_test", test_sqlite3rbu_internal_test },
   };
   int i;
-  for(i=0; i<sizeof(aObjCmd)/sizeof(aObjCmd[0]); i++){
+  for(i=0; i<(int)(sizeof(aObjCmd)/sizeof(aObjCmd[0])); i++){
     Tcl_CreateObjCommand(interp, aObjCmd[i].zName, aObjCmd[i].xProc, 0, 0);
   }
   return TCL_OK;
