+++ src/test_multiplex.cpp
@@ -1362,7 +1362,7 @@ int Sqlitemultiplex_Init(Tcl_Interp *interp){
   };
   int i;
 
-  for(i=0; i<sizeof(aCmd)/sizeof(aCmd[0]); i++){
+  for(i=0; i<(int)(sizeof(aCmd)/sizeof(aCmd[0])); i++){
     Tcl_CreateObjCommand(interp, aCmd[i].zName, aCmd[i].xProc, 0, 0);
   }
