+++ src/test_tclvar.cpp
@@ -554,7 +554,7 @@ int Sqlitetesttclvar_Init(Tcl_Interp *interp){
      { "register_tclvar_module",   register_tclvar_module, 0 },
   };
   int i;
-  for(i=0; i<sizeof(aObjCmd)/sizeof(aObjCmd[0]); i++){
+  for(i=0; i<(int)(sizeof(aObjCmd)/sizeof(aObjCmd[0])); i++){
     Tcl_CreateObjCommand(interp, aObjCmd[i].zName, 
         aObjCmd[i].xProc, aObjCmd[i].clientData, 0);
   }
